requirements.txt
fastapi
uvicorn
python-multipart
google-genai
chromadb
networkx
opencv-python-headless
scikit-image
pytesseract
Pillow
sentence-transformers
spacy
faster-whisper
yt-dlp
requests
pydantic
python-dotenv
pytest
